+++ config/AirTableFieldConfig.ts
@@ -1,7 +1,7 @@
 import { AirDateTimeFormatter } from '../enum/AirDateTimeFormatter'
 import { AirFieldConfig } from './AirFieldConfig'
 import { ITableFieldConfig } from '../interface/decorators/ITableFieldConfig'
-import { AirConfig } from '../config/AirConfig'
+import { AirConfig } from './AirConfig'
 import { AirDictionaryArray } from '../model/extend/AirDictionaryArray'
 import {
   AirSortable, AirTableAlign, AirTableFixed, ClassConstructor,
