+++ model/AirRequest.ts
@@ -1,6 +1,6 @@
 import { AirEntity } from '../base/AirEntity'
 import { AirModel } from '../base/AirModel'
-import { Field } from '../decorator/Field'
+import { Field } from '../decorator'
 import { AirClassTransformer } from '../helper/AirClassTransformer'
 import { ClassConstructor } from '../type/AirType'
 import { AirSort } from './AirSort'
