+++ model/AirRequestPage.ts
@@ -1,5 +1,5 @@
 import { AirEntity } from '../base/AirEntity'
-import { Field } from '../decorator/Field'
+import { Field } from '../decorator'
 import { AirPage } from './AirPage'
 import { AirRequest } from './AirRequest'
